move source files to src folder, some code updates..
