+++ src/AdvancedInstallerHelper/AdvancedInstallerHelper.vcxproj
@@ -107,7 +107,7 @@
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">Create</PrecompiledHeader>
     </ClCompile>
-    <ClCompile Include="helper.c" />
+    <ClCompile Include="customactions.c" />
   </ItemGroup>
   <ItemGroup>
     <None Include="exports.def" />
