+++ src/AdvancedInstallerHelper/AdvancedInstallerHelper.vcxproj.filters
@@ -23,13 +23,13 @@
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="helper.c">
-      <Filter>Source Files</Filter>
-    </ClCompile>
     <ClCompile Include="stdafx.c">
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="dllmain.c">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="customactions.c">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
