+++ src/AdvancedInstallerHelper/dllmain.c
