+++ src/AdvancedInstallerHelper/stdafx.c
