+++ src/AdvancedInstallerHelper/stdafx.h
