+++ src/wufuc/hooks.h
