+++ src/wufuc/patternfind.c
