+++ src/wufuc/patternfind.h
