+++ src/wufuc_setup_bat/COPYING.txt
