+++ src/wufuc_setup_bat/install_wufuc.bat
